feat: style hub action items
